requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
python-multipart==0.0.9

# Templates & Static
jinja2==3.1.4
aiofiles==24.1.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.8.0
pydantic-settings==2.4.0

# Database
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
greenlet==3.0.3

# Production
httpx==0.27.0
